+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### NEXT
 
-- SCTP: Encode subchannels in SCTP messages ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
+- SCTP: Encode subchannels in SCTP messages to enable subchannels mechanism when using pipe transport ([PR #XXXX](https://github.com/versatica/mediasoup/pull/XXXX)).
 
 ### 3.21.2